fix: fix bugs in delete review functionality

+++ src/main/java/com/saqlain/First/Project/review/impl/ReviewServiceImpl.java
@@ -59,9 +59,16 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public boolean deleteReview(Long companyId, Long reviewId) {
-        if(companyService.getCompanyById(companyId) != null)
+        if(companyService.getCompanyById(companyId) != null && reviewRepository.existsById(reviewId))
         {
+            Review review = reviewRepository.findById(reviewId).orElse(null);
+          //  assert review != null;
+            Company company = review.getCompany();
+            company.getReviews().remove(review);
+            companyService.updateCompany(company,companyId);
             reviewRepository.deleteById(reviewId);
+
+
             return true;
         }
         else{
